Revert "add common count time"
