+++ src/views/chat/areaChat/message/ChatMessageItem.js
@@ -9,7 +9,6 @@ import { Avatar, Box, Typography } from '@mui/material';
 import MessageStatus from './MessageStatus';
 import { selectMsgLatestStatus } from '../../../../features/chat/messagesSlice'
 import { useSelector } from 'react-redux';
-import { DateDistanceCustom } from '../../../common/time/dateDiff';
 // ----------------------------------------------------------------------
 
 const RootStyle = styled('div')(({ theme }) => ({
@@ -45,6 +44,36 @@ const MessageImgStyle = styled('img')(({ theme }) => ({
 ChatMessageItem.propTypes = {
   onOpenLightbox: PropTypes.func,
 };
+
+var DateDiff = {
+
+  inDays: function (d1, d2) {
+    var t2 = d2.getTime();
+    var t1 = d1.getTime();
+
+    return Math.floor((t2 - t1) / (24 * 3600 * 1000));
+  },
+
+  inWeeks: function (d1, d2) {
+    var t2 = d2.getTime();
+    var t1 = d1.getTime();
+
+    return parseInt((t2 - t1) / (24 * 3600 * 1000 * 7));
+  },
+
+  inMonths: function (d1, d2) {
+    var d1Y = d1.getFullYear();
+    var d2Y = d2.getFullYear();
+    var d1M = d1.getMonth();
+    var d2M = d2.getMonth();
+
+    return (d2M + 12 * d2Y) - (d1M + 12 * d1Y);
+  },
+
+  inYears: function (d1, d2) {
+    return d2.getFullYear() - d1.getFullYear();
+  }
+}
 
 export default function ChatMessageItem({ message, onOpenLightbox }) {
   const [showHistoryTimeFlg, setShowHistoryTimeFlg] = React.useState(false)
@@ -131,7 +160,7 @@ export default function ChatMessageItem({ message, onOpenLightbox }) {
                   className={`init-history-date non-block`}
                 >
                   {(message.created_at && showHistoryTimeFlg) ? (
-                   DateDistanceCustom(message.created_at)
+                    (DateDiff.inDays(new Date(message.created_at), new Date()) < 3) ? formatDistanceToNowStrict(new Date(message.created_at), { addSuffix: true }) : format(new Date(message.created_at), 'MM-dd-yyyy')
                   ) : ''}
                 </InfoStyle>
               </Box>
